+++ examples/ecs_app_plugin.rs
@@ -1,6 +1,6 @@
 use std::env;
 
-use new_ecs::{elements::PPNetwork, network::PowerFlowResult, plugin::*, post_processing::PostProcessing};
+use ecs::{elements::PPNetwork, network::PowerFlowResult, plugin::*, post_processing::PostProcessing};
 use rustpower::{io::pandapower::*, prelude::*};
 
 #[macro_export]
