+++ examples/ecs_vs_oop.rs
@@ -1,7 +1,7 @@
 #![allow(deprecated)]
 use std::env;
 use nalgebra::ComplexField;
-use new_ecs::{elements::PPNetwork, network::*, plugin::default_app, post_processing::*};
+use ecs::{elements::PPNetwork, network::*, plugin::default_app, post_processing::*};
 use rustpower::{io::pandapower::*, prelude::*};
 
 /// A utility macro to measure the execution time of a code block.
